+++ custom_components/my_wallet/panel.py
@@ -162,7 +162,7 @@ async def async_setup_panel(hass):
         webcomponent_name="my-wallet-panel",
         sidebar_title="My Wallet",
         sidebar_icon="mdi:chart-timeline-variant",
-        module_url="/my_wallet_static/my-wallet-panel.js?v=1.10.0",
+        module_url="/my_wallet_static/my-wallet-panel.js?v=1.10.1",
         embed_iframe=False,
         require_admin=True,
     )